.cursor/skills/kspeu-db-lab/scripts/requirements.txt
pypdf>=4.0,<6
pglast>=4.0,<8
aiofiles>=24.1.0,<25
python-docx>=1.1.0,<2
